+++ pkg/client/message/message.go
@@ -18,8 +18,8 @@ type sMessage struct {
 }
 
 func NewMessage(pEnck, pEncd []byte) IMessage {
-	if len(pEnck) != asymmetric.CKEncSize {
-		panic(`len(pEnck) != asymmetric.CKEncSize`)
+	if len(pEnck) != asymmetric.CKEMCiphertextSize {
+		panic(`len(pEnck) != asymmetric.CKEMCiphertextSize`)
 	}
 	return &sMessage{
 		fEnck: pEnck,
@@ -29,7 +29,7 @@ func NewMessage(pEnck, pEncd []byte) IMessage {
 
 // Message can be created only with client module.
 func LoadMessage(pSize uint64, pMsg interface{}) (IMessage, error) {
-	kSize := uint64(asymmetric.CKEncSize)
+	kSize := uint64(asymmetric.CKEMCiphertextSize)
 	if kSize >= pSize {
 		return nil, ErrSizeMessageBytes
 	}